sync transactions one by one instead of as a list

+++ app/src/main/java/cm/aptoide/pt/billing/Merchant.java
@@ -2,15 +2,15 @@ package cm.aptoide.pt.billing;
 
 public class Merchant {
 
-  private final int id;
+  private final long id;
   private final String name;
 
-  public Merchant(int id, String name) {
+  public Merchant(long id, String name) {
     this.id = id;
     this.name = name;
   }
 
-  public int getId() {
+  public long getId() {
     return id;
   }
